+++ config.js
@@ -1,4 +1,4 @@
-// concise helper frontend for writing a rooseveltConfig.js file
+// concise helper frontend for writing a roosevelt.config.js file
 // allows require('roosevelt/config') syntax in apps
 module.exports = {
   ref: require('./lib/tools/configRef')